+++ packages/terra-core-docs/src/terra-dev-site/test/text/TextLineHeight.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import Text from '../../../Text';
+import Text from 'terra-text';
 import styles from './TextTestCommon.module.scss';
 
 const cx = classNames.bind(styles);