+++ packages/terra-core-docs/CHANGELOG.md
@@ -3,9 +3,9 @@
 ## Unreleased
 
 * Added
-
   * Migrate docs from terra-visually-hidden-text.
   * Migrate docs from terra-toolbar.
+  * Migrate docs from terra-text.
 
 ## 1.12.2 - (January 4, 2022)